requirements.txt
PyYAML==6.0
pandas==1.4.2
matplotlib==3.5.2
requests==2.25.1
scikit-learn==0.24.1
pytest == 7.1.1